Add new response entity for user all

+++ services/auth/internal/repository/postgres/models/user.go
@@ -20,7 +20,7 @@ type User struct {
 
 type UserList struct {
 	Users      []User
-	TotalCount int
+	TotalCount uint64
 }
 
 type UserUpdatePassword struct {
